+++ weblate/trans/templatetags/translations.py
@@ -1774,6 +1774,7 @@ def format_last_changes_content(
                 },
                 "ip_address": change.get_ip_address() if user.is_superuser else None,
                 "history_data": get_change_history_context(change),
+                "message": change.message,
             }
         )
     return {